+++ package.json
@@ -85,7 +85,7 @@
   },
   "dependencies": {
     "@chakra-ui/react": "^1.5.2",
-    "@d-i-t-a/reader": "^1.8.4",
+    "@d-i-t-a/reader": "^1.8.6",
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.3.0",
     "@nypl/design-system-react-components": "^0.21.2",
