+++ src/ui/SettingsButton.tsx
@@ -32,7 +32,6 @@ export default function SettingsCard({ navigator }: { navigator: Navigator }) {
   }
 
   const paginationValue = navigator.isScrolling ? 'scrolling' : 'paginated';
-  console.log('isScrolling', navigator.isScrolling);
 
   return (
     <>
